# openapi.yaml
openapi: 3.0.3
info:
  title: Aframp Pay API
  version: 0.1.0
  description: >
    Merchant-facing backend for Aframp Pay — Stellar-settled point-of-sale payments.
    Prose reference with frontend guidance lives in API.md; this spec is the
    machine-readable contract for generating typed clients.


    Money is always integer stroops (1 unit = 10,000,000 stroops), never a float.
servers:
  - url: http://127.0.0.1:3000
    description: Local development
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /signup:
    post:
      tags: [Auth]
      summary: Start signup — sends an OTP
      description: >
        Does NOT create the account. Validates the credentials, sends an OTP
        to phone_number, and returns a challenge; the user and merchant are
        only inserted once POST /verify-otp succeeds. Re-submitting the same
        email before it's ever verified is the resend path, not a conflict.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpChallengeResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: Email or phone already registered to a verified account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429':
          description: OTP resend cooldown or hourly cap hit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /login:
    post:
      tags: [Auth]
      summary: Authenticate — sends an OTP if the phone is verified
      description: >
        Returns 401 for both a wrong password and an unknown email — deliberately
        indistinguishable, so do not derive a "no such account" message from it.
        On success, a phone-verified account gets a fresh OTP challenge (same
        shape as /signup) instead of a session — two-factor on every login.
        Only an account with no phone on file (pre-OTP-rollout only) gets the
        session directly, same as before.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Either an OTP challenge (normal case) or a session (legacy account only)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/OtpChallengeResponse'
                  - $ref: '#/components/schemas/AuthResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429':
          description: OTP resend cooldown or hourly cap hit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /verify-otp:
    post:
      tags: [Auth]
      summary: Complete a signup or login challenge
      description: >
        The only endpoint that ever issues a session. For a signup challenge,
        the user and merchant are created transactionally here, not before.
        Sets the aframp_session HttpOnly cookie and echoes the same JWT in the
        body for API clients. Browsers should rely on the cookie and ignore the
        token field — storing it in localStorage exposes the session to XSS.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyOtpRequest' }
      responses:
        '200':
          description: Session issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          description: OTP_INVALID, OTP_EXPIRED, or OTP_LOCKED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '404':
          description: OTP_CHALLENGE_NOT_FOUND — unknown or already-consumed challenge_id
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '409':
          description: Email or phone was taken by someone else between signup and verification
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
...
  /me:
    get:
      tags: [Auth]
      summary: Current user profile
      description: >
        The JWT carries only ids, so call this after a page reload to render
        human-readable identity without forcing a re-login.
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /wallet/create:
    post:
      tags: [Wallet]
      summary: Generate a Stellar wallet
      description: >
        Generates a real ed25519 keypair; the private key is AES-256-GCM encrypted
        server-side and never returned. NOT idempotent — each call creates another
        wallet, and GET /wallet then returns the newest. Check GET /wallet first.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWalletRequest' }
      responses:
        '200':
          description: Wallet created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wallet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
  /wallet:
    get:
      tags: [Wallet]
      summary: Get the merchant's most recent wallet
      responses:
        '200':
          description: Wallet
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wallet' }
        '400':
          description: No wallet created yet — run onboarding
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /payment-requests:
    post:
      tags: [Payment requests]
      summary: Create a payment request
      description: >
        The core POS action. Returns a SEP-0007 URI to render as a QR code.
        sep7_uri is null for non-XLM assets (no cNGN issuer address configured yet).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePaymentRequest' }
      responses:
        '200':
          description: Payment request created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentRequest' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
    get:
      tags: [Payment requests]
      summary: List the merchant's payment requests
      description: Newest first, scoped to the authenticated merchant.
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Payment requests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PaymentRequest' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /payment-requests/{id}:
    get:
      tags: [Payment requests]
      summary: Read a payment request (public)
      description: >
        Deliberately unauthenticated so a customer's device can read a request
        before paying. Prefer GET /payment-requests/{id}/status for polling
        after the customer has submitted payment.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Payment request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /payment-requests/{id}/status:
    get:
      tags: [Payment requests]
      summary: Poll payment-request status (public)
      description: >
        Lightweight status-only endpoint for customer-side polling after a QR
        payment. Uses the same effective_status rules as the full object
        (overdue pending → expired). Returns Cache-Control: public, max-age=5.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Status payload
          headers:
            Cache-Control:
              schema: { type: string, example: 'public, max-age=5' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentRequestStatus' }
        '404': { $ref: '#/components/responses/NotFound' }
  /balance:
    get:
      tags: [Money]
      summary: Balances by asset
      description: Returns an empty array for a new merchant — not an error.
      responses:
        '200':
          description: Balances
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Balance' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /transactions:
    get:
      tags: [Money]
      summary: Detected incoming payments
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Payment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /withdraw:
    post:
      tags: [Money]
      summary: Withdraw to a Nigerian bank account
      description: >
        Debits the balance and calls Paystack Transfers. Payouts do not currently
        complete — Aframp's Paystack balance is unfunded, so live calls return 502.
        On failure the balance is automatically refunded and the withdrawal is
        recorded with status "failed" plus a failure_reason; nothing is lost.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWithdrawalRequest' }
      responses:
        '200':
          description: Withdrawal accepted by the provider
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Withdrawal' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '502':
          description: Payout provider failed; balance already refunded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /withdrawals:
    get:
      tags: [Money]
      summary: List withdrawals
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Withdrawals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Withdrawal' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /webhooks/termii:
    post:
      tags: [Webhooks]
      summary: Termii SMS delivery-status callback
      description: >
        Register this URL at https://termii.com/account/webhook/config — it is
        a single account-wide setting, not something passed per API call.
        Termii POSTs SMS delivery-status events (Delivered, Message Failed,
        Rejected, etc.) signed with X-Termii-Signature (HMAC-SHA512 over the
        raw request body, keyed by TERMII_API_KEY). This endpoint verifies the
        signature, logs the event, and always acks with 204 for any recognized
        or unrecognized event shape so Termii never retries a genuine delivery.
        Not for merchant or admin use.
      security: []
      parameters:
        - name: X-Termii-Signature
          in: header
          required: true
          description: HMAC-SHA512 of the raw request body, keyed by TERMII_API_KEY.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TermiiWebhookEvent' }
      responses:
        '204':
          description: Event acknowledged (signature valid)
        '403':
          description: Invalid webhook signature
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
...
  /admin/overview:
    get:
      tags: [Admin]
      summary: System-wide counts and status breakdowns
      description: Aggregated totals across every merchant. Requires admin JWT.
      security:
        - adminBearerAuth: []
        - adminCookieAuth: []
      responses:
        '200':
          description: Overview
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminOverview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/merchants:
    get:
      tags: [Admin]
      summary: All merchants
      security:
        - adminBearerAuth: []
        - adminCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AdminLimit'
      responses:
        '200':
          description: Merchants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminMerchantRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/users:
    get:
      tags: [Admin]
      summary: All users
      security:
        - adminBearerAuth: []
        - adminCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AdminLimit'
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminUserRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/wallets:
    get:
      tags: [Admin]
      summary: All wallets
      security:
        - adminBearerAuth: []
        - adminCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AdminLimit'
      responses:
        '200':
          description: Wallets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminWalletRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/transactions:
    get:
      tags: [Admin]
      summary: All detected payments
      security:
        - adminBearerAuth: []
        - adminCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AdminLimit'
      responses:
        '200':
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminTransactionRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/withdrawals:
    get:
      tags: [Admin]
      summary: All withdrawals
      security:
        - adminBearerAuth: []
        - adminCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AdminLimit'
      responses:
        '200':
          description: Withdrawals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminWithdrawalRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/payment-requests:
    get:
      tags: [Admin]
      summary: All payment requests
      security:
        - adminBearerAuth: []
        - adminCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AdminLimit'
      responses:
        '200':
          description: Payment requests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminPaymentRequestRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        HS256, 24h expiry. No refresh endpoint — a 401 means re-login. For API
        clients and scripts; browsers should use cookieAuth instead.
    cookieAuth:
      type: apiKey
      in: cookie
      name: aframp_session
      description: >
        HttpOnly session cookie set by /signup and /login, cleared by /logout.
        The preferred method for browsers — JS cannot read it, so an XSS on the
        page cannot steal the session. Never copy the token into localStorage.
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Same JWT as bearerAuth, but the token must have been issued to a user
        with `is_admin = true`. The flag is baked into the JWT at login — flip
        the DB flag and re-login to pick up the change.
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: aframp_session
      description: >
        Same session cookie, but the JWT must carry `is_admin = true`.
        Re-login after flipping `is_admin` in Postgres.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Max rows to return. Clamped server-side to 1-200.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    BadRequest:
      description: Validation failed, or the account has no merchant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Missing, malformed, or expired token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    UnsupportedMediaType:
      description: >
        POST/PUT request carries a body but `Content-Type` is not
        `application/json` (or a `+json` subtype)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiError:
      type: object
      required: [error, code]
      properties:
        error: { type: string, example: insufficient available balance }
        code:
          type: string
          example: INSUFFICIENT_BALANCE
          description: "Machine-readable error code. Full catalog: INVALID_PARAMETERS, INVALID_AMOUNT, INSUFFICIENT_BALANCE, UNSUPPORTED_ASSET, PAYOUT_FAILED, EMAIL_TAKEN, PHONE_TAKEN, INVALID_CREDENTIALS, USER_NOT_FOUND, MERCHANT_NOT_FOUND, WALLET_NOT_FOUND, PAYMENT_REQUEST_NOT_FOUND, FORBIDDEN, OTP_INVALID, OTP_EXPIRED, OTP_LOCKED, OTP_CHALLENGE_NOT_FOUND, TOO_MANY_REQUESTS, INTERNAL_ERROR"
    SignupRequest:
      type: object
      required: [email, password, name, phone_number]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string, minLength: 1 }
        phone_number:
          type: string
          description: Nigerian mobile, any common format (0801..., 801..., +234801..., 234801...); normalized to E.164 server-side.
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    VerifyOtpRequest:
      type: object
      required: [challenge_id, code]
      properties:
        challenge_id: { type: string, format: uuid }
        code: { type: string, description: 6-digit code. }
    OtpChallengeResponse:
      type: object
      required: [challenge_id, expires_in_secs]
      properties:
        challenge_id: { type: string, format: uuid }
        expires_in_secs: { type: integer, example: 600 }
    AuthResponse:
      type: object
      required: [token, user_id]
      properties:
        token: { type: string }
        user_id: { type: string, format: uuid }
        merchant_id:
          type: string
          format: uuid
          nullable: true
          description: Null for an account with no merchant; such accounts get 400 from merchant-scoped endpoints.
    Me:
      type: object
      required: [user_id, email, name, created_at]
      properties:
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        created_at: { type: string, format: date-time }
        merchant_id: { type: string, format: uuid, nullable: true }
        merchant_name: { type: string, nullable: true }
    CreateWalletRequest:
      type: object
      properties:
        network: { type: string, default: stellar }
    Wallet:
      type: object
      required: [id, merchant_id, address, network, created_at]
      properties:
        id: { type: string, format: uuid }
        merchant_id: { type: string, format: uuid }
        address:
          type: string
          description: Stellar ed25519 public key (G...).
          example: GDDTPSD7BWERBIKVYXJY4KMBVFCUKNGJB2CS3DWBUUO3IB2CV7BZ5WSR
        network: { type: string, example: stellar }
        created_at: { type: string, format: date-time }
    CreatePaymentRequest:
      type: object
      required: [amount_stroops]
      properties:
        amount_stroops:
          type: integer
          format: int64
          minimum: 1
          description: >
            Amount in stroops (1 unit = 10,000,000 stroops). Must be a JSON
            integer — floats and strings are rejected with INVALID_PARAMETERS
            on field amount_stroops.
          example: 25000000
        asset: { type: string, default: XLM }
        expires_in_secs:
          type: integer
          format: int64
          default: 900
          description: Clamped server-side to 60-86400.
    PaymentRequest:
      type: object
      required: [id, merchant_id, address, network, amount_stroops, asset, memo, status, expires_at, created_at]
      properties:
        id: { type: string, format: uuid }
        merchant_id: { type: string, format: uuid }
        address: { type: string, description: Stellar destination address to pay. }
        network: { type: string, example: stellar }
        amount_stroops: { type: integer, format: int64 }
        asset: { type: string, example: XLM }
        memo:
          type: string
          description: >
            Correlation key. A payment without this memo still credits the balance
            but leaves the request pending forever. Included in sep7_uri automatically.
        status:
          type: string
          enum: [pending, paid, expired]
          description: >
            'expired' is computed at read time from expires_at. A request that
            expires and is then paid still flips to 'paid'.
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        sep7_uri:
          type: string
          nullable: true
          description: >
            SEP-0007 payment URI to render as a QR code. Null for non-XLM assets —
            no cNGN issuer address is configured, and a guessed one would misdirect funds.
          example: web+stellar:pay?destination=GDDT...&amount=2.5000000&memo=1f97c93409172a7d&memo_type=MEMO_TEXT
    PaymentRequestStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, paid, expired]
          description: >
            Same effective_status as GET /payment-requests/{id} — overdue
            pending rows report as expired at read time.
        paid_at:
          type: string
          format: date-time
          nullable: true
          description: Present when status is paid (row updated_at at mark-paid).
    Balance:
      type: object
      required: [merchant_id, asset, available, pending, updated_at]
      properties:
        merchant_id: { type: string, format: uuid }
        asset: { type: string, example: XLM }
        available: { type: integer, format: int64, description: Withdrawable, in stroops. }
        pending:
          type: integer
          format: int64
          description: Detected but unconfirmed. Almost always 0 — no confirmation-depth threshold yet.
        updated_at: { type: string, format: date-time }
...